Fixed build script to not fail if Makefile doesn't exist

+++ iphone/build-phonegap.sh
@@ -4,7 +4,7 @@ PHONEGAP_LIB=$TARGET_BUILD_DIR/$CONTENTS_FOLDER_PATH/www/app.js
 APPNAME=`defaults read $SOURCE_ROOT/Info CFBundleDisplayName`
 
 cd $PROJECT_DIR/..
-rm Makefile
+[ -f Makefile ] && rm Makefile
 ./configure
 make iphone
 cp lib/iphone/phonegap-min.js $PHONEGAP_LIB
